refactor(handlers): extract shared.ts (ToolResult + textResult/errorResult) (Stage B)
The two universal result constructors (used 149/213x) + the ToolResult type
move to a dependency-free module so handler modules can import them without
the dispatcher. ToolResult re-exported for back-compat. 3672 tests, A1 zero-diff.
