+++ src/handlers/shared.ts
@@ -1,0 +1,21 @@
+/**
+ * Shared primitives for the intent handlers: the MCP tool-result type and its two constructors.
+ *
+ * `textResult`/`errorResult` are used by every handler, so they live in a dependency-free module
+ * (extracted from intent.ts, Stage B) that handler modules can import without pulling in the whole
+ * dispatcher. `ToolResult` is re-exported from intent.ts for back-compat (barrel-surface.test.ts).
+ */
+
+/** MCP tool call result */
+export interface ToolResult {
+  content: Array<{ type: 'text'; text: string }>;
+  isError?: boolean;
+}
+
+export function textResult(text: string): ToolResult {
+  return { content: [{ type: 'text', text }] };
+}
+
+export function errorResult(message: string): ToolResult {
+  return { content: [{ type: 'text', text: message }], isError: true };
+}
